ATV 22 e 23: Criar Policy para Aluno e aplicar regras de permissao (Admin cria/exclui, Professor edita)

+++ app/Http/Controllers/AlunoController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Aluno;
 use Illuminate\Http\Request;
-use App\Http\Requests\AlunoRequest;    
+use App\Http\Requests\AlunoRequest;
 
 class AlunoController extends Controller
 {
@@ -22,29 +22,36 @@ class AlunoController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Aluno::class);
         return view('alunos.create');
     }
 
-public function store(AlunoRequest $request)
-{
-    Aluno::create($request->validated());
-    return redirect('/alunos');
-}
+    public function store(AlunoRequest $request)
+    {
+        $this->authorize('create', Aluno::class);
+        Aluno::create($request->validated());
+        return redirect('/alunos');
+    }
+
     public function edit($id)
     {
         $aluno = Aluno::findOrFail($id);
+        $this->authorize('update', $aluno);
         return view('alunos.edit', compact('aluno'));
     }
 
-public function update(AlunoRequest $request, $id)
-{
-    $aluno = Aluno::findOrFail($id);
-    $aluno->update($request->validated());
-    return redirect('/alunos');
-}
+    public function update(AlunoRequest $request, $id)
+    {
+        $aluno = Aluno::findOrFail($id);
+        $this->authorize('update', $aluno);
+        $aluno->update($request->validated());
+        return redirect('/alunos');
+    }
 
     public function destroy($id)
     {
+        $aluno = Aluno::findOrFail($id);
+        $this->authorize('delete', $aluno);
         Aluno::destroy($id);
         return redirect('/alunos');
     }
